+++ src/Symfony/Component/Mailer/Bridge/Postmark/Tests/Transport/PostmarkTransportFactoryTest.php
@@ -1,0 +1,74 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Mailer\Bridge\Postmark\Tests\Transport;
+
+use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkApiTransport;
+use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkSmtpTransport;
+use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkTransportFactory;
+use Symfony\Component\Mailer\Tests\TransportFactoryTestCase;
+use Symfony\Component\Mailer\Transport\Dsn;
+use Symfony\Component\Mailer\Transport\TransportFactoryInterface;
+
+class PostmarkTransportFactoryTest extends TransportFactoryTestCase
+{
+    public function getFactory(): TransportFactoryInterface
+    {
+        return new PostmarkTransportFactory($this->getDispatcher(), $this->getClient(), $this->getLogger());
+    }
+
+    public function supportsProvider(): iterable
+    {
+        yield [
+            new Dsn('api', 'postmark'),
+            true,
+        ];
+
+        yield [
+            new Dsn('smtp', 'postmark'),
+            true,
+        ];
+
+        yield [
+            new Dsn('smtp', 'example.com'),
+            false,
+        ];
+    }
+
+    public function createProvider(): iterable
+    {
+        $dispatcher = $this->getDispatcher();
+        $logger = $this->getLogger();
+
+        yield [
+            new Dsn('api', 'postmark', self::USER),
+            new PostmarkApiTransport(self::USER, $this->getClient(), $dispatcher, $logger),
+        ];
+
+        yield [
+            new Dsn('smtp', 'postmark', self::USER),
+            new PostmarkSmtpTransport(self::USER, $dispatcher, $logger),
+        ];
+    }
+
+    public function unsupportedSchemeProvider(): iterable
+    {
+        yield [
+            new Dsn('foo', 'postmark', self::USER),
+            'The "foo" scheme is not supported for mailer "postmark". Supported schemes are: "api", "smtp".',
+        ];
+    }
+
+    public function incompleteDsnProvider(): iterable
+    {
+        yield [new Dsn('api', 'postmark')];
+    }
+}