+++ src/Symfony/Bundle/WebProfilerBundle/CHANGELOG.md
@@ -13,7 +13,7 @@ CHANGELOG
  * Added button to clear the ajax request tab
  * Deprecated the `ExceptionController::templateExists()` method
  * Deprecated the `TemplateManager::templateExists()` method
- * Deprecated the `ExceptionController` in favor of `ExceptionErrorController`
+ * Deprecated the `ExceptionController` in favor of `ExceptionPanelController`
  * Marked all classes of the WebProfilerBundle as internal
 
 4.3.0